+++ collector/pkg/component/analyzer/network/config.go
@@ -1,15 +1,16 @@
 package network
 
 const (
-	defaultRequestTimeout        = 15
+	defaultFdReuseTimeout        = 15
 	defaultNoResponseThreshold   = 120
 	defaultConnectTimeout        = 1
 	defaultResponseSlowThreshold = 500
 )
 
 type Config struct {
-	ConnectTimeout int `mapstructure:"connect_timeout"`
-	RequestTimeout int `mapstructure:"request_timeout"`
+	ConnectTimeout      int `mapstructure:"connect_timeout"`
+	FdReuseTimeout      int `mapstructure:"fd_reuse_timeout"`
+	NoResponseThreshold int `mapstructure:"no_response_threshold"`
 	// unit is ms
 	ResponseSlowThreshold int `mapstructure:"response_slow_threshold"`
 
@@ -21,13 +22,13 @@ type Config struct {
 	ProtocolParser      []string         `mapstructure:"protocol_parser"`
 	ProtocolConfigs     []ProtocolConfig `mapstructure:"protocol_config,omitempty"`
 	UrlClusteringMethod string           `mapstructure:"url_clustering_method"`
-	NoResponseThreshold int              `mapstructure:"no_response_threshold"`
 }
 
 func NewDefaultConfig() *Config {
 	return &Config{
 		ConnectTimeout:        100,
-		RequestTimeout:        60,
+		FdReuseTimeout:        15,
+		NoResponseThreshold:   120,
 		ResponseSlowThreshold: 500,
 		EnableConntrack:       true,
 		ConntrackMaxStateSize: 131072,
@@ -70,7 +71,6 @@ func NewDefaultConfig() *Config {
 			},
 		},
 		UrlClusteringMethod: "alphabet",
-		NoResponseThreshold: 120,
 	}
 }
 
@@ -90,11 +90,11 @@ func (cfg *Config) GetConnectTimeout() int {
 	}
 }
 
-func (cfg *Config) GetRequestTimeout() int {
-	if cfg.RequestTimeout > 0 {
-		return cfg.RequestTimeout
+func (cfg *Config) GetFdReuseTimeout() int {
+	if cfg.FdReuseTimeout > 0 {
+		return cfg.FdReuseTimeout
 	} else {
-		return defaultRequestTimeout
+		return defaultFdReuseTimeout
 	}
 }